Changed the return type to float

+++ Cryptopals/Cryptopals/HammingDistanceCalculator.cs
@@ -14,7 +14,7 @@ namespace Cryptopals
     /// <param name="originalText">First text input</param>
     /// <param name="newText">The text to change to</param>
     /// <returns>An integer calculation representing the hamming distance</returns>
-    public int CalculateDistance(byte[] originalText, byte[] newText)
+    public float CalculateDistance(byte[] originalText, byte[] newText)
     {
       if (originalText.Length != newText.Length)
         throw new Exceptions.UnequalLengthException();
